exec

+++ execution/ft_here_doc.c
@@ -3,20 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   ft_here_doc.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: anomourn <anomourn@student.42.fr>          +#+  +:+       +#+        */
+/*   By: aeid <aeid@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/11 15:39:17 by aeid              #+#    #+#             */
-/*   Updated: 2024/07/27 17:56:05 by anomourn         ###   ########.fr       */
+/*   Updated: 2024/07/27 17:34:54 by aeid             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 # include "../headers/minishell.h"
 
-static char	*get_variable(char *string, t_data *data)
+static char *get_variable(char *string, t_data *data)
 {
-	int		i;
-	int		j;
-	char	*variable;
+	int i;
+	int j;
+	char *variable;
 	
 	i = ft_strlen(string);
 	memory_allocator((void **)&variable, i - 2, data);
@@ -28,11 +28,11 @@ static char	*get_variable(char *string, t_data *data)
 	return (variable);
 }
 
-static void	check_expansion(char *buffer, int fd, t_list *env, t_data *data)
+static void check_expansion(char *buffer, int fd, t_list *env, t_data *data)
 {
-	int		i;
-	char	*path;
-	char	*variable;
+	int i;
+	char *path;
+	char *variable;
 
 	i = 0;
 	path = NULL;
@@ -52,17 +52,18 @@ static void	check_expansion(char *buffer, int fd, t_list *env, t_data *data)
 		write(fd, buffer, ft_strlen(buffer));
 }
 
-void	ft_heredoc(t_list *file, t_list *env, t_data *data)
+void ft_heredoc(t_list *file, t_list *env, t_data *data, int *redi_num)
 {
-	t_tkn_data	*tokendata;
-	int			p_errno;
-	char		*buffer;
-	int			fd;
+	t_tkn_data *tokendata;
+	int p_errno;
+	char *buffer;
+	int fd;
 
 	tokendata = (t_tkn_data *)file->content;
 	fd = open(".heredoc", O_CREAT | O_RDWR | O_TRUNC, 0644);
 	while (1)
 	{
+		write(0, "> ", 2);
 		buffer = get_next_line(0);
 		//handle this properly
 		if (ft_strncmp(tokendata->token, buffer, ft_strlen(tokendata->token)) == 0 || buffer == NULL)
@@ -76,14 +77,18 @@ void	ft_heredoc(t_list *file, t_list *env, t_data *data)
 	get_next_line(-1);
 	free(buffer);
 	close(fd);
-	fd = open(".heredoc", O_RDONLY);
-	if (fd == -1)
+	if ((*redi_num) - 1 == 0)
 	{
-		p_errno = errno;
-		unlink(".heredoc");
-		ft_dup2_error_manager(fd, p_errno);
+		fd = open(".heredoc", O_RDONLY);
+		if (fd == -1)
+		{
+			p_errno = errno;
+			unlink(".heredoc");
+			ft_dup2_error_manager(fd, p_errno);
+		}
+		dup2(fd, 0);
+		close(fd);
 	}
-	dup2(fd, 0);
-	close(fd);
+	(*redi_num)--;
 	unlink(".heredoc");
 }
